requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.15.2
idna==3.10
lyricsgenius==3.3.1
mysqlclient==2.2.7
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
